requirements.txt
feedparser==6.0.10
requests==2.31.0
openai==1.3.0
elevenlabs==0.2.26
google-cloud-texttospeech==2.14.1
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pytest==7.4.3
httpx==0.25.1
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-multipart==0.0.6 